+++ react-app/src/library/services/headset.ts
@@ -1,15 +1,17 @@
 import { Observable, Subject } from 'rxjs';
 import { VendorImplementation, ImplementationConfig } from './vendor-implementations/vendor-implementation';
+import CyberAcousticsService from './vendor-implementations/CyberAcoustics/CyberAcoustics';
 import PlantronicsService from './vendor-implementations/plantronics/plantronics';
 import SennheiserService from './vendor-implementations/sennheiser/sennheiser';
 import JabraService from './vendor-implementations/jabra/jabra';
 import JabraNativeService from './vendor-implementations/jabra/jabra-native/jabra-native';
 import YealinkService from './vendor-implementations/yealink/yealink';
+import VBetService from './vendor-implementations/vbet/vbet';
 import { CallInfo } from '../types/call-info';
 import { VendorEvent, HoldEventInfo, MutedEventInfo, EventInfoWithConversationId } from '../types/emitted-headset-events';
 import { WebHidPermissionRequest } from '..';
 import { ConsumedHeadsetEvents, HeadsetEvents, DeviceConnectionStatus } from '../types/consumed-headset-events';
-import { HeadsetState, HeadsetStateRecord } from '../types/headset-states';
+import { HeadsetState, HeadsetStateRecord, UpdateReasons } from '../types/headset-states';
 
 type StateProps = Partial<HeadsetState>;
 type StateCompareProps = { conversationId: string; state: StateProps };
@@ -24,9 +26,11 @@ export default class HeadsetService {
   jabra: VendorImplementation;
   sennheiser: VendorImplementation;
   yealink: VendorImplementation;
+  vbet:VendorImplementation;
+  cyberAcoustics: VendorImplementation;
   selectedImplementation: VendorImplementation;
   headsetEvents$: Observable<ConsumedHeadsetEvents>;
-  
+
   private headsetConversationStates: { [conversationId: string]: HeadsetStateRecord } = {};
   private _headsetEvents$: Subject<ConsumedHeadsetEvents>;
   private logger: any;
@@ -41,8 +45,10 @@ export default class HeadsetService {
     this.jabra = JabraService.getInstance({ logger: this.logger });
     this.sennheiser = SennheiserService.getInstance({ logger: this.logger });
     this.yealink = YealinkService.getInstance({ logger: this.logger });
+    this.vbet = VBetService.getInstance({ logger: this.logger });
+    this.cyberAcoustics = CyberAcousticsService.getInstance({ logger: this.logger });
 
-    [this.plantronics, this.jabra, this.jabraNative, this.sennheiser, this.yealink].forEach(implementation => this.subscribeToHeadsetEvents(implementation));
+    [this.plantronics, this.jabra, this.jabraNative, this.sennheiser, this.yealink, this.vbet, this.cyberAcoustics].forEach(implementation => this.subscribeToHeadsetEvents(implementation));
   }
 
   static getInstance (config: ImplementationConfig): HeadsetService {
@@ -59,7 +65,9 @@ export default class HeadsetService {
       this.plantronics,
       this.jabra,
       this.jabraNative,
-      this.yealink
+      this.yealink,
+      this.vbet,
+      this.cyberAcoustics
     ].filter((impl) => impl.isSupported());
 
     return implementations;
@@ -72,11 +80,13 @@ export default class HeadsetService {
     return !state || Object.entries(props.state).some(([key, value]) => state[key] !== value);
   }
 
-  private updateHeadsetState (props: StateCompareProps): boolean {
+  private updateHeadsetState (props: StateCompareProps, opts = { expectExistingConversation: true }): boolean {
     if (this.isDifferentState(props)) {
       const state = this.headsetConversationStates[props.conversationId];
       if (!state) {
-        this.logger.warn('updateHeadsetState has no existing state for provided conversationId.', { conversationId: props.conversationId });
+        if (opts.expectExistingConversation) {
+          this.logger.warn('updateHeadsetState has no existing state for provided conversationId.', { conversationId: props.conversationId });
+        }
         return false;
       }
       Object.assign(state, props.state);
@@ -86,16 +96,27 @@ export default class HeadsetService {
     return false;
   }
 
-  activeMicChange (newMicLabel: string): void {
+  deviceIsSupported (params: { micLabel: string }): boolean {
+    if (!params.micLabel) {
+      return false;
+    }
+
+    const implementation = this.implementations.find((implementation) => {
+      return implementation.deviceLabelMatchesVendor(params.micLabel);
+    });
+    return !!implementation;
+  }
+
+  activeMicChange (newMicLabel: string, changeReason?: UpdateReasons): void {
     if (newMicLabel) {
       const implementation = this.implementations.find((implementation) => implementation.deviceLabelMatchesVendor(newMicLabel));
       if (implementation) {
         this.changeImplementation(implementation, newMicLabel);
       } else if (this.selectedImplementation) {
-        this.clearSelectedImplementation();
+        this.clearSelectedImplementation(changeReason);
       }
     } else {
-      this.clearSelectedImplementation();
+      this.clearSelectedImplementation(changeReason);
     }
   }
 
@@ -184,7 +205,7 @@ export default class HeadsetService {
     }
   }
 
-  async rejectCall (conversationId: string): Promise<any> {
+  async rejectCall (conversationId: string, expectExistingConversation = true): Promise<any> {
     const implementation = this.getConnectedImpl();
     if (!implementation) {
       return;
@@ -194,7 +215,7 @@ export default class HeadsetService {
       ringing: false
     };
 
-    if (this.updateHeadsetState({ conversationId, state: expectedStatePostAction })) {
+    if (this.updateHeadsetState({ conversationId, state: expectedStatePostAction }, { expectExistingConversation })) {
       const headsetState = this.headsetConversationStates[conversationId];
       headsetState.removeTimer = this.setRemoveTimer(conversationId);
       return implementation.rejectCall(conversationId);
@@ -250,7 +271,7 @@ export default class HeadsetService {
     if (!implementation) {
       return;
     }
-    
+
     Object.values(this.headsetConversationStates).forEach((headsetState) => {
       if (!headsetState.removeTimer) {
         headsetState.removeTimer = this.setRemoveTimer(headsetState.conversationId);
@@ -278,6 +299,15 @@ export default class HeadsetService {
     return 'noVendor';
   }
 
+  resetHeadsetStateForCall (conversationId: string): Promise<any> {
+    const implementation = this.getConnectedImpl();
+    if (implementation) {
+      return implementation.resetHeadsetStateForCall(conversationId);
+    } else {
+      this.logger.info('No active implementation, headset state does not require a reset');
+    }
+  }
+
   private getConnectedImpl (): VendorImplementation {
     const impl = this.selectedImplementation;
     if (!impl || !impl.isConnected) {
@@ -298,12 +328,12 @@ export default class HeadsetService {
     implementation.on(HeadsetEvents.webHidPermissionRequested, this.handleWebHidPermissionRequested.bind(this));
   }
 
-  private clearSelectedImplementation (): void {
+  private clearSelectedImplementation (clearReason?: UpdateReasons): void {
     if (!this.selectedImplementation) {
       return;
     }
 
-    this.selectedImplementation.disconnect();
+    this.selectedImplementation.disconnect(clearReason);
     this._headsetEvents$.next({ event: HeadsetEvents.implementationChanged, payload: null });
     this.selectedImplementation = null;
     this.handleDeviceConnectionStatusChanged();