+++ react-app/src/library/services/vendor-implementations/jabra/jabra.test.ts
@@ -63,13 +63,9 @@ const initializeSdk = async (subject?: Subject<IDevice[]>) => {
 
 describe('JabraService', () => {
   let jabraService: JabraService;
-  Object.defineProperty(window.navigator, 'hid', {
-    get: () => ({
-      getDevices: () => {
-        return [];
-      },
-    }),
-  });
+  (window.navigator as any) = { ...(window.navigator as any), hid: {
+    getDevices: jest.fn().mockResolvedValue([])
+  } };
   Object.defineProperty(window.navigator, 'locks', { get: () => ({}) });
   (window as any).BroadcastChannel = BroadcastChannel;
   
@@ -163,6 +159,7 @@ describe('JabraService', () => {
     it('should connect with previouslyConnectedDevice', async () => {
       const statusChangeSpy = jest.spyOn(jabraService, 'changeConnectionStatus');
       jest.spyOn(jabraService, 'getPreviouslyConnectedDevice').mockResolvedValue(mockDevice2 as any);
+      jest.spyOn(jabraService, 'deviceHasPermissions').mockResolvedValue(true);
 
       const callControl = createMockCallControl(new Subject<ICallControlSignal>().asObservable());
       jest
@@ -173,8 +170,34 @@ describe('JabraService', () => {
           },
         } as any);
 
+
       await jabraService.connect(mockDevice2.name);
       expect(statusChangeSpy).toHaveBeenCalledWith({ isConnected: true, isConnecting: false });
+    });
+
+    it('should attempt to connect with previouslyConnectedDevice but timeout', async () => {
+      const statusChangeSpy = jest.spyOn(jabraService, 'changeConnectionStatus');
+      const resetHeadsetSpy = jest.spyOn(jabraService, 'resetHeadsetState');
+      const processEventsSpy = jest.spyOn(jabraService, '_processEvents');
+      const callControl = createMockCallControl(new Subject<ICallControlSignal>().asObservable());
+      jest
+        .spyOn(jabraService, 'createCallControlFactory')
+        .mockReturnValue({
+          createCallControl: async () => {
+            return callControl;
+          },
+        } as any);
+      jest.useFakeTimers();
+      jest.spyOn(jabraService,'deviceHasPermissions').mockResolvedValue(true);
+      await jabraService.connect(mockDevice2.name);
+
+      await flushPromises();
+
+      jest.advanceTimersByTime(15005);
+
+      expect(statusChangeSpy).toHaveBeenCalledWith({ isConnected: false, isConnecting: false });
+      expect(resetHeadsetSpy).not.toHaveBeenCalled();
+      expect(processEventsSpy).not.toHaveBeenCalled();
     });
 
     it('should connect with webhidRequest', async () => {
@@ -217,6 +240,24 @@ describe('JabraService', () => {
   });
 
   describe('processEvents', () => {
+    it('ignores offhook events when there is an active and pending call', async () => {
+      jabraService.callLock = true;
+      const deviceSignalsSubject = new Subject<ICallControlSignal>();
+
+      const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
+      jabraService.activeConversationId = 'activeCid';
+      jabraService.pendingConversationId = 'pendingCid';
+      jabraService._processEvents(callControl as any);
+
+      deviceSignalsSubject.next({ type: 32, value: true } as any);
+      expect(callControl.offHook).not.toHaveBeenCalled();
+      expect(callControl.ring).not.toHaveBeenCalled();
+
+      deviceSignalsSubject.next({ type: 32, value: false } as any);
+      expect(callControl.offHook).not.toHaveBeenCalled();
+      expect(callControl.ring).not.toHaveBeenCalled();
+    });
+
     it('properly handles answer call events passed in from headset', async () => {
       jabraService.callLock = true;
       const deviceAnsweredCallSpy = jest.spyOn(jabraService, 'deviceAnsweredCall'); // const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
@@ -409,6 +450,26 @@ describe('JabraService', () => {
       expect(deviceRejectedCallSpy).toHaveBeenCalledWith({ conversationId, name: 'REJECT_CALL' });
       expect(await callControl.releaseCallLock).toHaveBeenCalled();
       expect(jabraService.callLock).toBe(false);
+    });
+
+    it('should only clean up if no active call', async () => {
+      jabraService.callLock = true;
+      const conversationId = 'convoId1234';
+      jabraService.pendingConversationId = conversationId;
+      jabraService.activeConversationId = 'active id';
+
+      const deviceRejectedCallSpy = jest.spyOn(jabraService, 'deviceRejectedCall');
+
+      const deviceSignalsSubject = new Subject<ICallControlSignal>();
+
+      const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
+      jabraService._processEvents(callControl as any);
+
+      deviceSignalsSubject.next({ type: 65533, value: true } as any);
+      expect(callControl.ring).toHaveBeenCalledWith(false);
+      expect(deviceRejectedCallSpy).toHaveBeenCalledWith({ conversationId, name: 'REJECT_CALL' });
+      expect(callControl.releaseCallLock).not.toHaveBeenCalled();
+      expect(jabraService.callLock).toBe(true);
     });
 
     it('properly handles reject call events passed in from headset with a failed callLock release', async () => {
@@ -612,7 +673,7 @@ describe('JabraService', () => {
       expect(connectionSpy).toHaveBeenCalled();
     });
 
-    it('should properly handle errors while attempting to release callLock', () => {
+    it('should properly handle errors while attempting to release callLock', async () => {
       const deviceSignalsSubject = new Subject<ICallControlSignal>();
       const connectionSpy = jabraService['changeConnectionStatus'] = jest.fn();
 
@@ -629,7 +690,7 @@ describe('JabraService', () => {
         );
       });
       const infoLoggerSpy = jest.spyOn(jabraService.logger, 'info');
-      jabraService.disconnect();
+      await jabraService.disconnect();
       expect(infoLoggerSpy).toHaveBeenCalledWith(
         'Trying to release the call lock, but it is not held!'
       );
@@ -638,7 +699,7 @@ describe('JabraService', () => {
       expect(connectionSpy).toHaveBeenCalled();
     });
 
-    it('should properly handle errors while attempting to release callLock', () => {
+    it('should properly handle errors while attempting to release callLock', async () => {
       const deviceSignalsSubject = new Subject<ICallControlSignal>();
       const connectionSpy = jabraService['changeConnectionStatus'] = jest.fn();
 
@@ -652,7 +713,7 @@ describe('JabraService', () => {
         throw exceptionWithType('Something much worse', ErrorType.UNEXPECTED_ERROR);
       });
       const errorLoggerSpy = jest.spyOn(jabraService.logger, 'error');
-      jabraService.disconnect();
+      await jabraService.disconnect();
       expect(errorLoggerSpy).toHaveBeenCalledWith(
         ErrorType.UNEXPECTED_ERROR,
         'Something much worse'
@@ -660,6 +721,25 @@ describe('JabraService', () => {
       expect(callControl.releaseCallLock).toHaveBeenCalled();
       expect(jabraService.callLock).toBe(false);
       expect(connectionSpy).toHaveBeenCalled();
+    });
+
+    it('should nullify activeConversationId if one exists upon disconnect', async () => {
+      const deviceSignalsSubject = new Subject<ICallControlSignal>();
+      const connectionSpy = jabraService['changeConnectionStatus'] = jest.fn();
+
+      const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
+
+      jabraService.callControl = callControl as any;
+      jabraService.callLock = true;
+      jabraService.isConnected = true;
+      jabraService.activeConversationId = 'testId123';
+
+      await jabraService.disconnect();
+
+      expect(callControl.releaseCallLock).toHaveBeenCalled();
+      expect(jabraService.callLock).toBe(false);
+      expect(connectionSpy).toHaveBeenCalled();
+      expect(jabraService.activeConversationId).toBeNull();
     });
   });
 
@@ -909,15 +989,34 @@ describe('JabraService', () => {
       const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
 
       jabraService.callControl = callControl as any;
+      const resetStateSpy = jest.spyOn(jabraService, 'resetState');
 
       const infoLoggerSpy = jest.spyOn(jabraService.logger, 'info');
-      const resetStateSpy = jest.spyOn(jabraService, 'resetState');
       await jabraService.rejectCall();
       expect(infoLoggerSpy).toHaveBeenCalledWith(
         'Currently not in possession of the Call Lock; Cannot react to Device Actions'
       );
       expect(jabraService.callLock).toBe(false);
-      expect(resetStateSpy).toHaveBeenCalled();
+      expect(resetStateSpy).not.toHaveBeenCalled();
+    });
+
+    it('should not release call lock if there is another active call', async () => {
+      const deviceSignalsSubject = new Subject<ICallControlSignal>();
+
+      const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
+      jabraService.activeConversationId = 'acd';
+      jabraService.callLock = true;
+
+      jabraService.callControl = callControl as any;
+
+      const infoLoggerSpy = jest.spyOn(jabraService.logger, 'info');
+      infoLoggerSpy.mockReset();
+      const resetStateSpy = jest.spyOn(jabraService, 'resetState');
+      await jabraService.rejectCall();
+      expect(infoLoggerSpy).not.toHaveBeenCalled();
+      expect(jabraService.callLock).toBe(true);
+      expect(resetStateSpy).not.toHaveBeenCalled();
+      expect(callControl.releaseCallLock).not.toHaveBeenCalled();
     });
 
     it('properly handles flow when in possession of callLock', () => {
@@ -944,7 +1043,7 @@ describe('JabraService', () => {
       expect(resetStateSpy).toHaveBeenCalled();
     });
 
-    it('properly handles error unrelated callLock', () => {
+    it('properly handles error unrelated callLock', async () => {
       const deviceSignalsSubject = new Subject<ICallControlSignal>();
 
       const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
@@ -956,7 +1055,7 @@ describe('JabraService', () => {
         throw exceptionWithType('Something much worse', ErrorType.UNEXPECTED_ERROR);
       });
       const errorLoggerSpy = jest.spyOn(jabraService.logger, 'error');
-      jabraService.rejectCall();
+      await jabraService.rejectCall();
       expect(errorLoggerSpy).toHaveBeenCalledWith(
         ErrorType.UNEXPECTED_ERROR,
         'Something much worse'
@@ -1002,13 +1101,11 @@ describe('JabraService', () => {
       jabraService.callControl = callControl as any;
 
       const infoLoggerSpy = jest.spyOn(jabraService.logger, 'info');
-      const resetStateSpy = jest.spyOn(jabraService, 'resetState');
       await jabraService.endCall('123', false);
       expect(infoLoggerSpy).toHaveBeenCalledWith(
         'Currently not in possession of the Call Lock; Cannot react to Device Actions'
       );
       expect(jabraService.callLock).toBe(false);
-      expect(resetStateSpy).toHaveBeenCalled();
     });
 
     it('properly handles flow when in possession of callLock', () => {
@@ -1035,7 +1132,7 @@ describe('JabraService', () => {
       expect(resetStateSpy).toHaveBeenCalled();
     });
 
-    it('properly handles error unrelated callLock', () => {
+    it('properly handles error unrelated callLock', async () => {
       const deviceSignalsSubject = new Subject<ICallControlSignal>();
 
       const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
@@ -1047,7 +1144,7 @@ describe('JabraService', () => {
         throw exceptionWithType('Something much worse', ErrorType.UNEXPECTED_ERROR);
       });
       const errorLoggerSpy = jest.spyOn(jabraService.logger, 'error');
-      jabraService.endCall('123', false);
+      await jabraService.endCall('123', false);
       expect(errorLoggerSpy).toHaveBeenCalledWith(
         ErrorType.UNEXPECTED_ERROR,
         'Something much worse'
@@ -1080,13 +1177,11 @@ describe('JabraService', () => {
       jabraService.callControl = callControl as any;
 
       const infoLoggerSpy = jest.spyOn(jabraService.logger, 'info');
-      const resetStateSpy = jest.spyOn(jabraService, 'resetState');
       await jabraService.endAllCalls();
       expect(infoLoggerSpy).toHaveBeenCalledWith(
         'Currently not in possession of the Call Lock; Cannot react to Device Actions'
       );
       expect(jabraService.callLock).toBe(false);
-      expect(resetStateSpy).toHaveBeenCalled();
     });
 
     it('properly handles flow when in possession of callLock', () => {
@@ -1113,7 +1208,7 @@ describe('JabraService', () => {
       expect(resetStateSpy).toHaveBeenCalled();
     });
 
-    it('properly handles error unrelated callLock', () => {
+    it('properly handles error unrelated callLock', async () => {
       const deviceSignalsSubject = new Subject<ICallControlSignal>();
 
       const callControl = createMockCallControl(deviceSignalsSubject.asObservable());
@@ -1125,7 +1220,7 @@ describe('JabraService', () => {
         throw exceptionWithType('Something much worse', ErrorType.UNEXPECTED_ERROR);
       });
       const errorLoggerSpy = jest.spyOn(jabraService.logger, 'error');
-      jabraService.endAllCalls();
+      await jabraService.endAllCalls();
       expect(errorLoggerSpy).toHaveBeenCalledWith(
         ErrorType.UNEXPECTED_ERROR,
         'Something much worse'
@@ -1181,6 +1276,14 @@ describe('JabraService', () => {
     it('should return false if no deviceInfo', () => {
       jabraService._deviceInfo = null;
       expect(jabraService.isDeviceAttached).toEqual(false);
+    });
+  });
+
+  describe('resetHeadsetStateForCall', () => {
+    it('should call the rejectCall function', () => {
+      const rejectSpy = jest.spyOn(jabraService, 'rejectCall');
+      jabraService.resetHeadsetStateForCall();
+      expect(rejectSpy).toHaveBeenCalled();
     });
   });
 
@@ -1299,7 +1402,21 @@ describe('JabraService', () => {
       await expect(devicePromise).rejects.toThrow('random error');
     });
   });
-  
+
+  describe('deviceHasPermissions', () => {
+    Object.defineProperty(window.navigator, 'hid', {
+      get: () => ({
+        getDevices: () => { return [{ productName: 'test-device' } as any]; }
+      })
+    });
+    it('should return true if passed in label exists in getDevices', async () => {
+      const deviceHasPermissions = await jabraService.deviceHasPermissions('test-device');
+      await flushPromises();
+
+      expect(deviceHasPermissions).toBe(true);
+    });
+  });
+
   describe('getPreviouslyConnectedDevice', () => {
     afterEach(() => {
       jest.useRealTimers();
@@ -1343,22 +1460,6 @@ describe('JabraService', () => {
 
       const device = await devicePromise;
       expect(device).toBeNull();
-    });
-
-    it('should timeout after 3 seconds', async () => {
-      jest.useFakeTimers();
-
-      const sub = new BehaviorSubject([]);
-      jabraService.jabraSdk = await initializeSdk(sub as any) as any;
-
-      const devicePromise = jabraService.getPreviouslyConnectedDevice(mockDevice2.name);
-
-      await flushPromises();
-
-      jest.advanceTimersByTime(3100);
-
-      const device = await devicePromise;
-      expect(device).toBeFalsy();
     });
 
     it('should log random error', async () => {