+++ react-app/src/library/services/vendor-implementations/jabra/jabra.ts
@@ -91,6 +91,12 @@ export default class JabraService extends VendorImplementation {
 
       switch (signal.type) {
       case SignalType.HOOK_SWITCH:
+        // do nothing when the end call button is pressed and we have an incoming call while we have an active call
+        if (this.activeConversationId && this.pendingConversationId) {
+          this.logger.info('ignoring hookswitch event because there is an active and incoming call');
+          break;
+        }
+
         if (signal.value) {
           callControl.offHook(true);
           callControl.ring(false);
@@ -153,17 +159,19 @@ export default class JabraService extends VendorImplementation {
           name: SignalType[signal.type],
           conversationId: this.pendingConversationId,
         });
+        this.pendingConversationId = null;
         try {
-          callControl.releaseCallLock();
+          // we only want to release call controls if there isn't another call active
+          if (!this.activeConversationId) {
+            callControl.releaseCallLock();
+            this.callLock = false;
+          }
         } catch ({ message, type }) {
           if (this.checkForCallLockError(message, type)) {
             this.logger.info(message);
           } else {
             this.logger.error(type, message);
           }
-        } finally {
-          this.pendingConversationId = null;
-          this.callLock = false;
         }
       }
     });
@@ -230,24 +238,27 @@ export default class JabraService extends VendorImplementation {
   }
 
   async rejectCall (): Promise<void> {
-    try {
-      if (!this.callLock) {
-        return this.logger.info(
-          'Currently not in possession of the Call Lock; Cannot react to Device Actions'
-        );
+    if (!this.callLock) {
+      return this.logger.info(
+        'Currently not in possession of the Call Lock; Cannot react to Device Actions'
+      );
+    }
+    this.callControl.ring(false);
+    this.pendingConversationId = null;
+    if (!this.activeConversationId) {
+      try {
+        this.resetState();
+        this.callControl.releaseCallLock();
+      } catch ({ message, type }) {
+        if (this.checkForCallLockError(message, type)) {
+          this.logger.info(message);
+        } else {
+          this.logger.error(type, message);
+        }
+      } finally {
+        this.pendingConversationId = null;
+        this.callLock = false;
       }
-      this.callControl.ring(false);
-      this.callControl.releaseCallLock();
-    } catch ({ message, type }) {
-      if (this.checkForCallLockError(message, type)) {
-        this.logger.info(message);
-      } else {
-        this.logger.error(type, message);
-      }
-    } finally {
-      this.pendingConversationId = null;
-      this.callLock = false;
-      this.resetState();
     }
   }
 
@@ -286,6 +297,7 @@ export default class JabraService extends VendorImplementation {
         );
       }
       this.callControl.offHook(false);
+      this.resetState();
       this.callControl.releaseCallLock();
     } catch ({ message, type }) {
       if (this.checkForCallLockError(message, type)) {
@@ -295,7 +307,6 @@ export default class JabraService extends VendorImplementation {
       }
     } finally {
       this.callLock = false;
-      this.resetState();
     }
   }
 
@@ -308,6 +319,7 @@ export default class JabraService extends VendorImplementation {
       }
       this.activeConversationId = null;
       this.callControl.offHook(false);
+      this.resetState();
       this.callControl.releaseCallLock();
     } catch ({ message, type }) {
       if (this.checkForCallLockError(message, type)) {
@@ -317,7 +329,6 @@ export default class JabraService extends VendorImplementation {
       }
     } finally {
       this.callLock = false;
-      this.resetState();
     }
   }
 
@@ -340,8 +351,16 @@ export default class JabraService extends VendorImplementation {
 
     this._deviceInfo = null;
 
-    let selectedDevice = await this.getPreviouslyConnectedDevice(deviceLabel);
-    if (!selectedDevice) {
+    let selectedDevice;
+    if (await this.deviceHasPermissions(deviceLabel)) {
+      selectedDevice = await this.getPreviouslyConnectedDevice(deviceLabel);
+
+      if (!selectedDevice) {
+        console.warn('Unable to find appropriate device. Setting state to "Not Running" to allow a retry"', deviceLabel);
+        this.changeConnectionStatus({ isConnected: false, isConnecting: false });
+        return;
+      }
+    } else {
       try {
         selectedDevice = await this.getDeviceFromWebhid(deviceLabel);
       } catch (e) {
@@ -363,8 +382,18 @@ export default class JabraService extends VendorImplementation {
     this.changeConnectionStatus({ isConnected: true, isConnecting: false });
   }
 
+  async deviceHasPermissions (deviceLabel: string): Promise<boolean> {
+    const allowedHIDDevices = await (window.navigator as any).hid.getDevices();
+    let deviceFound = false;
+    allowedHIDDevices.forEach(device => {
+      if (deviceLabel.includes(device?.productName?.toLowerCase())) {
+        deviceFound = true;
+      }
+    });
+    return deviceFound;
+  }
+
   async getPreviouslyConnectedDevice (deviceLabel: string): Promise<IDevice> {
-    // we want to wait for up to 2 events or timeout after 2 seconds
     const waitForDevice: Observable<IDevice> = this.jabraSdk.deviceList.pipe(
       defaultIfEmpty(null),
       first((devices: IDevice[]) => !!devices.length),
@@ -372,7 +401,7 @@ export default class JabraService extends VendorImplementation {
         devices.find((device) => this.isDeviceInList(device, deviceLabel))
       ),
       filter((device) => !!device),
-      timeout(3000)
+      timeout(15000)
     );
 
     return firstValueFrom(waitForDevice).catch((err) => {
@@ -430,11 +459,15 @@ export default class JabraService extends VendorImplementation {
       return;
     }
 
-    await this.callControl.takeCallLock();
-    this.callControl.hold(false);
-    this.callControl.mute(false);
-    this.callControl.offHook(false);
-    this.callControl.releaseCallLock();
+    try {
+      await this.callControl.takeCallLock();
+      this.callControl.hold(false);
+      this.callControl.mute(false);
+      this.callControl.offHook(false);
+      this.callControl.releaseCallLock();
+    } catch (e) {
+      this.logger.warn('Failed to takeCallLock in order to resetHeadsetState. Ignoring reset.');
+    }
   }
 
   async disconnect (): Promise<void> {
@@ -452,7 +485,11 @@ export default class JabraService extends VendorImplementation {
         this.logger.error(type, message);
       }
     } finally {
+      this.resetHeadsetState();
       this.callLock = false;
+      if (this.activeConversationId) {
+        this.activeConversationId = null;
+      }
       this.headsetEventSubscription && this.headsetEventSubscription.unsubscribe();
       (this.isConnected || this.isConnecting) &&
         this.changeConnectionStatus({ isConnected: false, isConnecting: false });