+++ react-app/src/library/services/vendor-implementations/vendor-implemntation.test.ts
@@ -113,3 +113,24 @@ describe('isSupported', () => {
     expect(implementation.isSupported()).toBeTruthy();
   });
 });
+
+describe('resetHeadsetStateForCall', () => {
+  it('should call rejectCall', async () => {
+    const rejectSpy = implementation.rejectCall = jest.fn();
+    await implementation.resetHeadsetStateForCall('test123');
+    expect(rejectSpy).toHaveBeenCalledWith('test123');
+  });
+});
+
+describe('deductProductId', () => {
+  it('should match proper labels', () => {
+    const result = implementation.deductProductId('Test Device Label (6993:b017)');
+    expect(result).toBe(45079);
+  });
+
+  it('should not match invalid labels', () => {
+    const result = implementation.deductProductId('Test Device Label');
+    expect(result).toBe(null);
+  });
+});
+